Fix AngularJS issues
When opening a Person form, we don't display anymore marketing data.
Reason is that we renamed `meta` to `marketing` but forgot to update JS code.

Also, `person.id` is shaded with `@JsonIgnore` annotation so in JS code we need to use `person.reference` instead.

Closes #8.
(cherry picked from commit 29c36a0)